Separate load and generate files
